Users objects ypes

See merge request blockchain-consultora/itsmine-backend!6

+++ functions/src/controllers/chats.controller.ts
@@ -1,10 +1,10 @@
-import GlobalReq from '../database/global.request';
-import { fieldsSchema, postSchema, patchSchema } from '../models/object.models';
+import ChatRequest from '../database/chats.request';
+import { fieldsSchema, postSchema, patchSchema } from '../models/chats.models';
 import { Request, Response } from "express";
 import { db, myBucket } from "../database/connect";
 import storage from "../storage/index";
 
-const objects = new GlobalReq('objects', fieldsSchema);
+const objects = new ChatRequest('chats', fieldsSchema);
 
 export const singleGet = async (req: Request, res: Response) => {
   try {
@@ -22,10 +22,22 @@ export const singleGet = async (req: Request, res: Response) => {
 
 export const countMethod = async (req: Request, res: Response) => {
   try {
+    const size = await objects.countMyChats(storage.getUserId());
+    return res.json({ count: size });
+  } catch (err) {
+    let msg = err.message;
+
+    return res.status(400).json({ msg });
+  }
+}
+
+export const getMethodOwn = async (req: Request, res: Response) => {
+  try {
     let filter = typeof req.query.filter === 'string' && JSON.parse(req.query.filter);
 
-    const size = await objects.countDocuments(filter);
-    return res.json({ count: size });
+    const arrRes = await objects.getMyChats(filter, storage.getUserId());
+
+    return res.json(arrRes);
   } catch (err) {
     let msg = err.message;
 
@@ -50,8 +62,9 @@ export const getMethod = async (req: Request, res: Response) => {
 export const postMethod = async (req: Request, res: Response) => {
   try {
     let value = await postSchema.validateAsync(req.body);
+    let timestamps = { createdAt: Date.now(), lastMessage: Date.now() }
 
-    let obj = await objects.addDocument({ ...value, userId: storage.getUserId() });
+    let obj = await objects.addDocument({ ...value, ...timestamps });
     return res.json(obj);
 
   } catch (err) {
@@ -81,34 +94,16 @@ export const patchMethod = async (req: Request, res: Response) => {
 export const deleteMethod = async (req: Request, res: Response) => {
   try {
     if (!req.params.id) throw new Error('id is required');
+    let { members } = await objects.getSingleDoc(req.params.id);
 
-    await objects.deleteDocument(req.params.id);
+    if (members.length == 1)
+      await objects.deleteDocument(req.params.id);
+    else {
+      let restMem: Array<string> = members.filter((ele: any) => ele != storage.getUserId());
 
-    return res.json({ msg: 'success deleted' });
-  } catch (err) {
-    let msg = err.message;
-
-    return res.status(400).json({ msg });
-  }
-}
-
-export const deleteFilesMethod = async (req: Request, res: Response) => {
-  try {
-    if (!req.params.id) throw new Error('id is required');
-    let { images } = await objects.getSingleDoc(req.params.id);
-
-    let arLink = images[0].link.split('/');
-    let leng = arLink.length;
-
-    let path = `images/${arLink[leng - 2]}/`
-
-    await myBucket.deleteFiles({ prefix: path });
-
-    for (let el of images) {
-      await db.collection("uploadFiles").doc(el.id).delete();
+      await objects.setDocument(req.params.id, { members: restMem });
     }
 
-    await objects.deleteDocument(req.params.id);
 
     return res.json({ msg: 'success deleted' });
   } catch (err) {
@@ -117,3 +112,29 @@ export const deleteFilesMethod = async (req: Request, res: Response) => {
     return res.status(400).json({ msg });
   }
 }
+
+// export const deleteFilesMethod = async (req: Request, res: Response) => {
+//   try {
+//     if (!req.params.id) throw new Error('id is required');
+//     let { images } = await objects.getSingleDoc(req.params.id);
+
+//     let arLink = images[0].link.split('/');
+//     let leng = arLink.length;
+
+//     let path = `images/${arLink[leng - 2]}/`
+
+//     await myBucket.deleteFiles({ prefix: path });
+
+//     for (let el of images) {
+//       await db.collection("uploadFiles").doc(el.id).delete();
+//     }
+
+//     await objects.deleteDocument(req.params.id);
+
+//     return res.json({ msg: 'success deleted' });
+//   } catch (err) {
+//     let msg = err.message;
+
+//     return res.status(400).json({ msg });
+//   }
+// }
