+++ functions/src/controllers/notification.controller.ts
@@ -1,16 +1,15 @@
 import GlobalReq from '../database/global.request';
-import { fieldsSchema, postSchema, patchSchema } from '../models/object.models';
+import { fieldsSchema, postSchema, patchSchema } from '../models/notification.models';
 import { Request, Response } from "express";
-import { db, myBucket } from "../database/connect";
-import storage from "../storage/index";
+import { db, myBucket } from "../database/connect"
 
-const objects = new GlobalReq('objects', fieldsSchema);
+const notifications = new GlobalReq('notifications', fieldsSchema);
 
 export const singleGet = async (req: Request, res: Response) => {
   try {
     if (!req.params.id) throw new Error('id is required');
 
-    let doc = await objects.getSingleDoc(req.params.id);
+    let doc = await notifications.getSingleDoc(req.params.id);
 
     return res.json(doc);
   } catch (err) {
@@ -24,7 +23,7 @@ export const countMethod = async (req: Request, res: Response) => {
   try {
     let filter = typeof req.query.filter === 'string' && JSON.parse(req.query.filter);
 
-    const size = await objects.countDocuments(filter);
+    const size = await notifications.countDocuments(filter);
     return res.json({ count: size });
   } catch (err) {
     let msg = err.message;
@@ -37,7 +36,7 @@ export const getMethod = async (req: Request, res: Response) => {
   try {
     let filter = typeof req.query.filter === 'string' && JSON.parse(req.query.filter);
 
-    const arrRes = await objects.getCollection(filter);
+    const arrRes = await notifications.getCollection(filter);
 
     return res.json(arrRes);
   } catch (err) {
@@ -51,9 +50,8 @@ export const postMethod = async (req: Request, res: Response) => {
   try {
     let value = await postSchema.validateAsync(req.body);
 
-    let obj = await objects.addDocument({ ...value, userId: storage.getUserId() });
+    let obj = await notifications.addDocument(value);
     return res.json(obj);
-
   } catch (err) {
     console.log(err);
     let msg = err.details ? err.details : err.msg
@@ -66,8 +64,8 @@ export const postMethod = async (req: Request, res: Response) => {
 export const patchMethod = async (req: Request, res: Response) => {
   try {
     let value = await patchSchema.validateAsync(req.body);
-    let obj = await objects.setDocument(req.params.id, value);
 
+    let obj = await notifications.setDocument(req.params.id, value);
     return res.json(obj);
   } catch (err) {
     console.log(err);
@@ -82,7 +80,7 @@ export const deleteMethod = async (req: Request, res: Response) => {
   try {
     if (!req.params.id) throw new Error('id is required');
 
-    await objects.deleteDocument(req.params.id);
+    await notifications.deleteDocument(req.params.id);
 
     return res.json({ msg: 'success deleted' });
   } catch (err) {
@@ -92,23 +90,22 @@ export const deleteMethod = async (req: Request, res: Response) => {
   }
 }
 
-export const deleteFilesMethod = async (req: Request, res: Response) => {
+export const deleteFileMethod = async (req: Request, res: Response) => {
   try {
     if (!req.params.id) throw new Error('id is required');
-    let { images } = await objects.getSingleDoc(req.params.id);
+    let { avatar } = await notifications.getSingleDoc(req.params.id);
 
-    let arLink = images[0].link.split('/');
-    let leng = arLink.length;
+    if (avatar) {
+      let arLink = avatar.link.split('/');
+      let leng = arLink.length;
 
-    let path = `images/${arLink[leng - 2]}/`
+      let path = `images/${arLink[leng - 2]}/`
 
-    await myBucket.deleteFiles({ prefix: path });
-
-    for (let el of images) {
-      await db.collection("uploadFiles").doc(el.id).delete();
+      await myBucket.deleteFiles({ prefix: path });
+      await db.collection("uploadFiles").doc(avatar.id).delete();
     }
 
-    await objects.deleteDocument(req.params.id);
+    await notifications.deleteDocument(req.params.id);
 
     return res.json({ msg: 'success deleted' });
   } catch (err) {
